Refactored issue_retrieval.py
~Consolidated both issues and prs into the same data retrieval calls

-Eliminated any function pass-through

+Modified .json to accept both comments and openings in different sections
