+++ .gitignore
@@ -503,18 +503,6 @@ dmypy.json
 *mies
 
 # Other names that would make sense
-*tests
-*testsdir
-*testsfile
-*testsfiles
-*test
-*testdir
-*testfile
-*testfiles
-*testing
-*testingdir
-*testingfile
-*testingfiles
 *temp
 *tempdir
 *tempfile
