--- data/demofile_2.json
@@ -1,81 +1,0 @@
-{
-    "stephensonc": {
-        "commits": [
-            "commit hash 1",
-            "commit hash 2",
-            "commit hash 3"
-        ],
-        "issues": [
-            4,
-            4,
-            4,
-            4,
-            4
-        ],
-        "pull_requests": [
-            23,
-            23,
-            23
-        ]
-    },
-    "koscinskic": {
-        "commits": [
-            "commit hash 1",
-            "commit hash 2",
-            "commit hash 3"
-        ],
-        "issues": [
-            30,
-            30,
-            29,
-            24,
-            20,
-            15,
-            15,
-            15,
-            15,
-            15,
-            14,
-            14,
-            14,
-            9,
-            8,
-            8,
-            7,
-            7,
-            4,
-            1
-        ],
-        "pull_requests": [
-            6,
-            6,
-            6
-        ]
-    },
-    "schultzh": {
-        "commits": [
-            "commit hash 1",
-            "commit hash 2",
-            "commit hash 3"
-        ],
-        "issues": [
-            30,
-            30,
-            30,
-            29,
-            27,
-            26,
-            24,
-            16,
-            16,
-            12,
-            10
-        ],
-        "pull_requests": [
-            31,
-            17,
-            6,
-            6
-        ]
-    }
-}
